fix(build): transform i18n import from dynamic to static
