+++ packages/brisa/src/brisa-project-internals.ts
@@ -1,7 +1,11 @@
 import importFileIfExists from '@/utils/import-file-if-exists';
-import { getConstants } from '@/constants';
+import { internalConstants } from '@/utils/load-constants';
+import type { I18nConfig } from "brisa";
 
-const { BUILD_DIR } = getConstants();
+// All this is temporal to work in DEV (brisa dev), in the future DEV is going to work
+// in the same way that PROD works (without dynamic imports and these constants would
+// be needed only in build-time, inside attach-brisa-project-internals)
+const { BUILD_DIR, WORKSPACE } = internalConstants();
 
 /**
  * WIP https://github.com/brisa-build/brisa/issues/628
@@ -12,16 +16,17 @@ const { BUILD_DIR } = getConstants();
  * Useful to avoid dynamic imports. Being possible to compile the app into binary to
  * improve memory in runtime.
  */
-const middlewareModule = await importFileIfExists('middleware', BUILD_DIR);
-export const middleware = middlewareModule?.default;
+export const middleware = (await importFileIfExists('middleware', BUILD_DIR))
+  ?.default;
+export const i18n = (await importFileIfExists('i18n', WORKSPACE))
+  ?.default as I18nConfig;
 
 // TODO:
-export const i18n = null;
 export const config = null;
+export const integrations = null;
+export const cssFiles = [];
 export const api = [];
 export const layout = null;
 export const websockets = null;
 export const actions = null;
 export const pages = [];
-export const integrations = null;
-export const cssFiles = [];
